+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -2,6 +2,8 @@ import { config } from "~/libs/modules/config/config.js";
 import { database } from "~/libs/modules/database/database.js";
 import { logger } from "~/libs/modules/logger/logger.js";
 import { authController } from "~/modules/auth/auth.js";
+import { onboardingController } from "~/modules/onboarding/onboarding.js";
+import { quizController } from "~/modules/quiz/quiz.js";
 import { userController } from "~/modules/users/users.js";
 
 import { BaseServerApplication } from "./base-server-application.js";
@@ -12,6 +14,8 @@ const apiV1 = new BaseServerApplicationApi(
 	config,
 	...authController.routes,
 	...userController.routes,
+	...quizController.routes,
+	...onboardingController.routes,
 );
 const serverApplication = new BaseServerApplication({
 	apis: [apiV1],