+++ apps/backend/src/modules/quiz-answers/libs/types/types.ts
@@ -1,0 +1,10 @@
+export { type CategorizedQuizAnswerModel } from "./categorized-quiz-answer-model.type.js";
+export { type CategoryStatistic } from "./category-statistic.type.js";
+export { type UserAnswersRequestDto } from "./user-answers-request-dto.type.js";
+export {
+	type QuizAnswerDto,
+	type QuizAnswerRequestDto,
+	type QuizAnswersRequestDto,
+	type QuizAnswersResponseDto,
+	type QuizUserAnswerDto,
+} from "shared";