+++ packages/shared/src/libs/enums/error-message.enum.ts
@@ -1,5 +1,8 @@
 const ErrorMessage = {
+	DUPLICATE_QUESTION_ANSWER: "There are multiple answers to the same question.",
 	INCORRECT_CREDENTIALS: "Incorrect credentials.",
+	INSUFFICIENT_ANSWERS: "You must provide answers for all questions.",
+	REQUESTED_ENTITY_NOT_FOUND: "The requested entity was not found.",
 	UNAUTHORIZED: "You are unauthorized to access the requested resource.",
 } as const;