+++ packages/shared/src/modules/onboarding/libs/types/onboarding-answer-request-dto.type.ts
@@ -1,0 +1,6 @@
+type OnboardingAnswerRequestDto = {
+	answerIds: number[];
+	userId: number;
+};
+
+export { type OnboardingAnswerRequestDto };