+++ packages/shared/src/modules/quiz/libs/types/quiz-answer-request-dto.type.ts
@@ -1,0 +1,7 @@
+type QuizAnswerRequestDto = {
+	label: string;
+	questionId: number;
+	value: number;
+};
+
+export { type QuizAnswerRequestDto };